requirements.txt
streamlit
pandas
altair
cassandra-driver
